+++ apps/web/partials/review/bounty-linking/index.ts
@@ -7,3 +7,5 @@ export {
   isBountyTypeRelation,
   isAllocatedToUser,
 } from './build-bounties';
+export { useLinkableBounties } from './use-linkable-bounties';
+export { usePublishBountyLinks } from './use-publish-bounty-links';
